Add sound randomization + cleanup weapon code
Instead of playing the same soudn effect each time, we can now randomly
cycle between multiple sounds. All of the sound fields are tables instead of strings.
Instead of pulling the string of a sound file from the table field and passing it
directly to minetest.sound_play(), we get a random value from the table
of sounds with boomstick.get_random_sound(), which uses boomstick.get_random_entry().